feat: Removes token processing module from `phoenix.evals` (#2421)
* Remove token counting functionality from `phoenix.evals`

* Deprecate `phoenix.experimental.evals.functions.processing`
